Assign state program types
All state program types should now be correct, except for Arizona.  AZ will need to be changed to "combo" but not yet. It's currently the only way we can test the "separate_chip" functionality.  We'll want to make that change after we get fixtures in place for WA or CT.