Edge Case: Not Combination State
If a State was either Medicaid Expansion or Separate CHIP, the numbers were loading as null for the other program and causing the page to not display.

At some point in the future, we will need to handle removing or excluding the row if the State does not have that program type.