Use state list when creating new state statuses.
When the admin user triggers generation of a new form year, the
StateStatus objects were only being created for States that had
a prior StateStatus object. This would result in no StateStatus
objects being created in a brand new environment, like local.

Here, we switch is enumerating over the States themselves to ensure
that a StateStatus object exists for every state for the new form
year.