Remove ses password notice service
Looking at invocations, this service has only been erroring and relates
only to remainders to rotate the Database User's password. These remainders
are currently handled in JIRA, and we are moving off of postgres soon.

Instead of fixing up this service, removing it. If we stay on postgres,
then suggests we do automated password rotations instead of a password
change remainder email.