Hardcode the name of the SEDS state-forms topic
This value is not secret, and it does not change.

Having an environment variable for the topic name is useful in theory:
it allows us to control the topic environment-by-environment.
If we make some change to how we consume Kafka data,
we will be able to control that change (partially) by flipping the variable.

But only partially.
For example, if the new topic has some other data format,
we would need to change the code which parses the data.
And if we wanted to feature flag that code change,
we would need to create a new feature flag,
regardless of this environment variable.

Having the topic name hardcoded makes the system easier to understand.
A developer who wants to know which Kafka topic CARTS subscribes to
will not need to dig through CDK code to find out how the env var is set.
They won't need to log in to dev AWS to check the value of the secret JSON.
They won't need to wonder if the secret differs from env to env.
They won't need admin permissions to production AWS to verify the value there.
In short, they won't need to do all the work I had to do a few weeks ago.

If we do need to change this variable,
that will be a significant project
requiring coordination between SEDS, CARTS, and DataConnect.
The presence or absence of this environment variable
will not make a significant difference
to the effort required for such an undertaking.